+++ src/main/java/com/kurkus/kusinsa/service/product/ProductStatusService.java
@@ -2,7 +2,7 @@ package com.kurkus.kusinsa.service.product;
 
 import com.kurkus.kusinsa.dto.request.product.status.ProductStockUpdateRequest;
 import com.kurkus.kusinsa.entity.Product;
-import com.kurkus.kusinsa.repository.ProductRepository;
+import com.kurkus.kusinsa.repository.product.ProductRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
