+++ src/main/resources/application-dev.properties
@@ -11,7 +11,7 @@ spring.data.mongodb.database=kusinsa
 spring.jpa.hibernate.ddl-auto=validate
 spring.jpa.generate-ddl=false
 spring.jpa.show-sql=true
-#spring.jpa.properties.hibernate.format_sql=true
+spring.jpa.properties.hibernate.format_sql=true
 logging.level.org.hibernate.SQL=debug
 # redis session
 spring.redis.session.host=localhost
