+++ src/test/java/com/kurkus/kusinsa/service/order/OrderHistoryServiceTest.java
@@ -7,7 +7,7 @@ import com.kurkus.kusinsa.entity.Product;
 import com.kurkus.kusinsa.entity.User;
 import com.kurkus.kusinsa.enums.DeliveryStatus;
 import com.kurkus.kusinsa.repository.OrderHistoryRepository;
-import com.kurkus.kusinsa.repository.ProductRepository;
+import com.kurkus.kusinsa.repository.product.ProductRepository;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
