+++ src/test/java/com/kurkus/kusinsa/service/product/ProductServiceTest.java
@@ -16,8 +16,7 @@ import com.kurkus.kusinsa.exception.product.ProductException;
 import com.kurkus.kusinsa.exception.product.ProductNotFoundException;
 import com.kurkus.kusinsa.repository.BrandRepository;
 import com.kurkus.kusinsa.repository.CategoryRepository;
-import com.kurkus.kusinsa.repository.ProductRepository;
-import com.kurkus.kusinsa.service.product.ProductService;
+import com.kurkus.kusinsa.repository.product.ProductRepository;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
