+++ src/test/java/com/kurkus/kusinsa/test.java
@@ -1,30 +1,17 @@
 package com.kurkus.kusinsa;
 
 
-import java.time.LocalDateTime;
-import java.util.*;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 
-import com.kurkus.kusinsa.dao.LikesDao;
-import com.kurkus.kusinsa.dao.RankDao;
 import com.kurkus.kusinsa.dto.request.notification.NotificationCreateRequest;
-import com.kurkus.kusinsa.dto.response.rank.OrderRankResponse;
-import com.kurkus.kusinsa.entity.Product;
-import com.kurkus.kusinsa.entity.documents.ClickRank;
-import com.kurkus.kusinsa.enums.notification.NotificationGroupStatus;
-import com.kurkus.kusinsa.repository.ProductRepository;
 import com.kurkus.kusinsa.repository.UserRepository;
-import com.kurkus.kusinsa.repository.mongo.ClickRankRepository;
 import com.kurkus.kusinsa.repository.notification.NotificationGroupRepository;
-import com.kurkus.kusinsa.service.RankService;
 import com.kurkus.kusinsa.service.notification.NotificationService;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.data.redis.core.RedisTemplate;
-import org.springframework.transaction.annotation.Transactional;
 
 @SpringBootTest
 public class test {
